# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Production backend for UnityWorks Vision AI. Owns the Vision OS perception platform and exposes it through an authorized API.",
    "title": "UnityWorks Vision AI",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/users": {
      "get": {
        "description": "Users in the caller's organization, searchable and filterable.\n\nThe role filter is a subquery rather than a join, deliberately: a user may\nhold several roles, and joining would return them once per matching role\nand make `total` wrong in a way that only shows up for exactly the accounts\nan administrator is most likely to be looking at.",
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Users Api V1 Admin Users Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Users",
        "tags": [
          "user-administration"
        ]
      },
      "post": {
        "description": "Create a user in the caller's organization.\n\nA duplicate email is a clean `ConflictError` (409), never a 500. Email is\nunique across the deployment (`app/users/models.py:uq_users_email`), so the\naddress may already belong to somebody at another customer \u2014 and saying so\nwould confirm to one customer that a person has an account at another. The\nrefusal therefore says only that the address cannot be used here and that a\nPlatform Admin can give an existing account access. Where it is held goes\nto the audit trail, which is for the people entitled to know.\n\nEvery role in `roles` must be grantable by the actor (rule 2 in the module\ndocstring, `_require_grantable_role`); an unknown role name is a\n`ValidationError` (422).\n\n### `camera_scope` is required, and that is the fix for a real bug\n\nThis route used to create a `User` and its `RoleAssignment` rows and stop.\nNo `AccessGrant` was written, and a missing grant reads as\n`CameraScope.none()` \u2014 correctly, since that is the only safe reading of an\nabsent row. The result was an account that logged in, held every permission\nits role carried, and could reach no cameras at all. Nothing reported it.\n`admin1@unityworks.local` is in exactly that state in production today.\n\nThe fix is not a default. Both candidate defaults are wrong: `none`\nsilently recreates the unusable account, and `all_in_tenant` silently\ncreates an over-privileged one. So the caller states which of the three\nscopes they mean, and `none` remains a legitimate answer \u2014 for an account\nthat will be scoped later, or one that never needs video.",
        "operationId": "create_user_api_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create User Api V1 Admin Users Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User",
        "tags": [
          "user-administration"
        ]
      }
    },
...
    "/api/v1/zones": {
      "get": {
        "description": "Zones in the caller's organization, with how many cameras each holds.\n\nGated on `VIEW_ZONES`: somebody reading incidents needs to know the places\nthey are attributed to, and reading grants no ability to change them.",
        "operationId": "list_zones_api_v1_zones_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Zones Api V1 Zones Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Zones",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "description": "Create a zone: a name, and nothing else to decide.\n\nEverything a zone needs is its name \u2014 kitchen, dine hall, billing. The\norganization comes from the session, never from the request.",
        "operationId": "create_zone_api_v1_zones_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Zone Api V1 Zones Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Zone",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/zones/{zone_id}": {
      "delete": {
        "description": "Delete a zone that holds no cameras.\n\nA zone with cameras is refused with the count and the alternative, because\nit is part of the record of what was watched and where. Deactivating keeps\nthat record; deleting would lose it.",
        "operationId": "delete_zone_api_v1_zones__zone_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Zone Api V1 Zones  Zone Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Zone",
        "tags": [
          "administration"
        ]
      },
      "get": {
        "operationId": "get_zone_api_v1_zones__zone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Zone Api V1 Zones  Zone Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Zone",
        "tags": [
          "administration"
        ]
      },
      "patch": {
        "description": "Rename a zone, or take it out of use.\n\nRenaming does not rewrite history: `CameraZoneAssignment` froze the name a\npast reading was labelled with, deliberately.",
        "operationId": "update_zone_api_v1_zones__zone_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "zone_id",
            "required": true,
            "schema": {
              "title": "Zone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Zone Api V1 Zones  Zone Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Zone",
        "tags": [
          "administration"
        ]
      }
    },
...
  }
}